Hadoop/python-container/requirements.txt
pandas
psycopg2-binary
hdfs
happybase
